CU Instances list, ServiceOrder entity editing

+++ InternetProvider/InternetProvider-web/src/main/java/com/netcracker/wind/commands/implementations/order/ProceedToOrder.java
@@ -81,7 +81,7 @@ public class ProceedToOrder implements ICommand {
         orderDAO.add(order);
         try {
             jsono.put("auth", true);
-            jsono.put("order_id", order.getId());
+            jsono.put("order_id", order.getId().intValue());
         } catch (JSONException ex) {
             Logger.getLogger(ProceedToOrder.class.getName()).log(Level.SEVERE, null, ex);
         }
